+++ Integration Tests/DepotsTests/UsersDepotTests.cs
@@ -1,4 +1,6 @@
-﻿using CSM_Security_Database_Core.Depots;
+﻿using CSM_Database_Core.Depots.Models;
+
+using CSM_Security_Database_Core.Depots;
 using CSM_Security_Database_Core.Entities;
 
 using CSM_Security_Database_Testing.Abstractions.Bases;
@@ -15,12 +17,32 @@ public class UsersDepotTests
     protected override User EntityFactory(string Entropy) {
         return DraftUtils.User(
                 new User {
-                    UserInfo = Store(DraftUtils.UserInfo())
+                    UserInfo = Store(DraftUtils.UserInfo()).GetAwaiter().GetResult(),
                 }
             );
     }
 
-    public override Task Update_Single_Success() {
-        throw new NotImplementedException();
+    public override async Task Update_Single_Success() {
+        //Preparing
+        User user = await _storeManager.StoreUser();
+        
+        //Expectations
+        string oldUsername = user.Username;
+        string newUsername = $"{oldUsername}_upd";
+        
+        //Acting
+        user.Username = newUsername;
+        UpdateOutput<User> actOutput = await _depot.Update(
+                new QueryInput<User, UpdateInput<User>> {
+                    Parameters = new UpdateInput<User> {
+                        Entity = user,
+                    }
+                }
+            );
+
+        //Asserting
+        Assert.NotNull(actOutput.Original);
+        Assert.Equal(oldUsername, actOutput.Original.Username);
+        Assert.Equal(newUsername, actOutput.Updated.Username);
     }
 }